+++ app/src/main/java/woowacourse/movie/ui/history/BookingHistoryPresenter.kt
@@ -1,11 +1,10 @@
-package woowacourse.movie.ui.history.presenter
+package woowacourse.movie.ui.history
 
+import kotlin.concurrent.thread
 import woowacourse.movie.data.BookedTicketDao
 import woowacourse.movie.data.BookedTicketDatabase
 import woowacourse.movie.data.BookedTicketEntity
 import woowacourse.movie.data.toBookedTicket
-import woowacourse.movie.ui.history.contract.BookingHistoryContract
-import kotlin.concurrent.thread
 
 class BookingHistoryPresenter(
     private val view: BookingHistoryContract.View,
